+++ joblib/__init__.py
@@ -124,8 +124,7 @@ from .parallel import register_parallel_backend
 from .parallel import parallel_backend
 from .parallel import effective_n_jobs
 
-from .externals.loky import wrap_non_picklable_objects
-
+from .externals.cloudpickle_wrapper import wrap_non_picklable_objects
 
 __all__ = ['Memory', 'MemorizedResult', 'PrintTime', 'Logger', 'hash', 'dump',
            'load', 'Parallel', 'delayed', 'cpu_count', 'effective_n_jobs',
