+++ joblib/externals/cloudpickle_wrapper.py
